requirements.txt
PyQt5==5.15.11
python-dotenv==1.0.1
langchain-groq==0.2.2
langchain-core==0.3.29
